requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
distro==1.9.0
fastapi==0.135.1
groq==1.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ngrok==1.4.0
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
pytesseract==0.3.13
python-dotenv==1.2.2
python-multipart==0.0.22
sniffio==1.3.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
